requirements.txt
python-telegram-bot>=20.5
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=1.0
